+++ source/themes/blocks.js
@@ -14,15 +14,15 @@ newtheme.buildgrid = function(gridcontainer){
   var hasleftorrightsection = false;
   var footerline;
   var sections = new Array();
-  console.log("build grid for "+gridcontainer.classList);
+  //console.log("build grid for "+gridcontainer.classList);
   for(var x=0;x<gridelemente.length;x++){
   	var gridel = gridelemente[x];
-    console.log("grid: start with element:"+gridel.innerHTML);
-    console.log("grid: element has clientHeight:"+gridel.clientHeight);
+    //console.log("grid: start with element:"+gridel.innerHTML);
+    //console.log("grid: element has clientHeight:"+gridel.clientHeight);
     //console.log("gridarea before change:"+gridel.style.gridArea);
     if(gridel.classList.contains("bgimg")){
       //background-image found - dont insert into gridarray
-      console.log("grid: backgroundimage found");
+      //console.log("grid: backgroundimage found");
       continue;
     }
   	if(!gridel.style.gridArea){
@@ -32,14 +32,14 @@ newtheme.buildgrid = function(gridcontainer){
     if(gridel.classList.contains("section")&&
         (gridel.classList.contains("left") ||gridel.classList.contains("center") || gridel.classList.contains("right"))){
       //maxheight = window.innerHeight;
-      console.log("left, center or right found:");
+      //console.log("left, center or right found:");
       //save grid-y to object to attach left-section afterwards
       gridel.gridy = gridy;
       sections.push(gridel); //save left or right-section to sectionsarray
       hasleftorrightsection = true;
       continue; //do nothing more - dont attach left-right-section to array yet
     }
-    console.log("grid: again clientHeight"+gridel.clientHeight);
+    //console.log("grid: again clientHeight"+gridel.clientHeight);
     //console.log("gridel-gridarea:"+gridel.style.gridArea);
   	if(gridel.classList.contains("vertical")){
   	 //element is vertical:
@@ -51,7 +51,7 @@ newtheme.buildgrid = function(gridcontainer){
       if(gridx===1 && heightleft>0 && maxheight>heightleft
           && this.clientHeight(gridel, "60%")>heightleft){
         //element does not fit next to neighbour:
-        console.log("grid: vertical does not fit into column - go to next row");
+        //console.log("grid: vertical does not fit into column - go to next row");
         //gridx=0;gridy++;
         //gridarray[gridy]=new Array();
         gridx=0;
@@ -59,8 +59,8 @@ newtheme.buildgrid = function(gridcontainer){
   		gridarray[gridy][gridx]=gridel;
   		maxheight=this.clientHeight(gridel) *1.5;
       heightleft = maxheight;
-      console.log("grid: vertical found with maxheight:"+maxheight);
-      console.log("grid: clientHeight:"+gridel.clientHeight);
+      //console.log("grid: vertical found with maxheight:"+maxheight);
+      //console.log("grid: clientHeight:"+gridel.clientHeight);
   		gridx++;
   		if(gridx>2){
   			gridx=0;gridy++;
@@ -68,7 +68,7 @@ newtheme.buildgrid = function(gridcontainer){
   		}
   	}else if(gridel.classList.contains("horizontal")){
   	 //element is horizontal - always break:
-     console.log("grid:horizontal element found");
+     //console.log("grid:horizontal element found");
   		if(gridx>0){
         gridx=0;
         if(heightleft===maxheight)gridy++; //vertical starts in same row
@@ -81,7 +81,7 @@ newtheme.buildgrid = function(gridcontainer){
   		//element is flexible:
   		if(gridx==0){
   			//easiest case: just print it into the whole row:
-        console.log("grid: add element easy to row"+gridy +" with clientHeight"+gridel.clientHeight);
+        //console.log("grid: add element easy to row"+gridy +" with clientHeight"+gridel.clientHeight);
   			gridarray[gridy]=[gridel];
   			gridy++;
   		}else if(gridx==1){
@@ -90,12 +90,12 @@ newtheme.buildgrid = function(gridcontainer){
   			//gridel.classList.add("block-flex-neighbour"); //add class to get width to 60%;
   			var height=this.clientHeight(gridel, "60%");//gridel.clientHeight;	//get actual height of element if put as neighbour
         //console.log("grid: height of element after:"+height);
-        console.log("grid: height:"+height+"heightleft:"+heightleft);
+        //console.log("grid: height:"+height+"heightleft:"+heightleft);
   			//gridel.classList.remove("block-flex-neighbour"); //remove class to not mess with other css
   			if(heightleft<height){
   				//element does not fit into column, check if put into next row or use this row:
   				if(heightleft<maxheight){
-            console.log("grid: heightleft<maxheight - add element to row"+gridy);
+            //console.log("grid: heightleft<maxheight - add element to row"+gridy);
   					//vertical element has one neighbour yet, so put it in the same row, deleting old vertical in this row
   					gridx=0;
   					gridarray[gridy]=[gridel];
@@ -104,7 +104,7 @@ newtheme.buildgrid = function(gridcontainer){
   					maxheight=0;
   				}else{
   					//switch to next row:
-            console.log("grid: switch to next row after row:"+gridy);
+            //console.log("grid: switch to next row after row:"+gridy);
   					gridx=0;
   					gridy++;
   					gridarray[gridy]=[gridel];
@@ -117,7 +117,7 @@ newtheme.buildgrid = function(gridcontainer){
   				gridarray[gridy][gridx]=gridel;
   				//if more elements are coming continue with vertical in column 1
           if(gridelemente.length > x+1){
-            console.log("gridelementelength:"+gridelemente.length+"x:"+x);
+            //console.log("gridelementelength:"+gridelemente.length+"x:"+x);
             gridy++;
             gridarray[gridy]=new Array();
             gridarray[gridy][0]=gridarray[gridy-1][0];
@@ -137,17 +137,17 @@ newtheme.buildgrid = function(gridcontainer){
 
   }
   //check for left-right-sections:
-  console.log("has left or right section:"+hasleftorrightsection);
+  //console.log("has left or right section:"+hasleftorrightsection);
   //hasleftorrightsection = false;
   if(hasleftorrightsection){
     if(gridarray.length>1){
-      console.log("check for footer:"+gridarray[gridarray.length-1][0].tagName);
+      //console.log("check for footer:"+gridarray[gridarray.length-1][0].tagName);
       if(gridarray[gridarray.length-1][0].tagName==="FOOTER"){
         footerline = gridarray.pop();
       }
     }
     //check if no content after left/right-sections:
-    console.log("gridarray.length:"+gridarray.length+"sections[0].gridy:"+sections[0].gridy);
+    //console.log("gridarray.length:"+gridarray.length+"sections[0].gridy:"+sections[0].gridy);
     if(gridarray.length<=sections[0].gridy){
       gridarray.push(new Array());
     }
@@ -200,10 +200,10 @@ newtheme.buildgrid = function(gridcontainer){
   	//do not only build 2-colums-template if neighbour is not vertical - build a 3-column instead:
   	if(gridarray[x].length ==2 && colums ==2 && !gridarray[x][1].classList.contains("vertical"))colums=3;
   }
-  console.log("gridarray:");
-  console.log(gridarray);
+  //console.log("gridarray:");
+  //console.log(gridarray);
   //build lastline for leftright-sections:
-  console.log(gridarray);
+  //console.log(gridarray);
   //now build the area-array:
   var gridarea = new Array();
   for(var x=0;x<gridarray.length;x++){
@@ -233,8 +233,8 @@ newtheme.buildgrid = function(gridcontainer){
       }
     }
   }//end of for-to-sections.length
-  console.log("gridarea:");
-  console.log(gridarea);
+  //console.log("gridarea:");
+  //console.log(gridarea);
   //the area-array is now filled with the right values - build an area out of it:
   var area = '';
   for(var x=0;x<gridarea.length;x++){
@@ -246,7 +246,7 @@ newtheme.buildgrid = function(gridcontainer){
   area+='" ';
   for(var y=0;y<colums;y++)area+="nav "
   area+='"';
-  console.log("area to use:"+area);
+  //console.log("area to use:"+area);
   gridcontainer.style.gridTemplateAreas = area;
   //gridcontainer.style.gridTemplateRows = "repeat("+gridarray.length+", 1fr )";
   gridcontainer.classList.add("gridx"+colums);
@@ -278,7 +278,7 @@ newtheme.buildgrid = function(gridcontainer){
   let plholder = document.createElement("div");
   plholder.className = "navigationPlaceholder";
   gridcontainer.appendChild(plholder);
-  console.log("gridTemplateRows:"+rowheights);
+  //console.log("gridTemplateRows:"+rowheights);
   gridcontainer.style.gridTemplateRows=rowheights;
   gridcontainer.classList.add("gridy"+(rows+1));
 }
@@ -301,7 +301,7 @@ newtheme.clientHeight = function(element, width){
   slidenote.presentationdiv.appendChild(el);
   if(width)el.style.width = width;
   var result = el.clientHeight;
-  console.log("grid: clientHeight returned"+result);
+  //console.log("grid: clientHeight returned"+result);
   //document.getElementsByTagName("body")[0].removeChild(el);
   slidenote.presentationdiv.removeChild(el);
   return result;
@@ -321,7 +321,7 @@ newtheme.addBlockClassesToElements = function(gridcontainer){
   var nodes = gridcontainer.children;
   for(var e=0;e<nodes.length;e++){
     var node = nodes[e];
-    console.log("grid: addClass to Node:"+node.className+"/"+node.nodeName);
+    //console.log("grid: addClass to Node:"+node.className+"/"+node.nodeName);
     if(node.className === "imageblock" || node.nodeName==="IMG"){
       //imageblock: get clientWidth and clientHeight;
       var cliw = this.clientWidth(node); //does not work like this:
@@ -332,7 +332,7 @@ newtheme.addBlockClassesToElements = function(gridcontainer){
       //console.log("grid: imagesinblock[0].naturalWidth:"+imagesinblock[0].naturalWidth);
       var clih = this.clientHeight(node, cliw);
       var landscape = (cliw - clih > 0);
-      console.log("grid: imagew:"+cliw+"imageh:"+clih);
+      //console.log("grid: imagew:"+cliw+"imageh:"+clih);
       if(!landscape)node.classList.add("vertical");
       if(landscape)node.classList.add("horizontal");
       if(nodes.length==1){
@@ -344,24 +344,24 @@ newtheme.addBlockClassesToElements = function(gridcontainer){
       //  node.classList.add("bgimg");
       //}
     }// end of imageblock
-    console.log("nodeName: "+node.nodeName);
+    //console.log("nodeName: "+node.nodeName);
     if(node.nodeName === "OL" || node.nodeName === "UL"){
-      console.log("grid: list found clientwidth:"+node.clientWidth);
+      //console.log("grid: list found clientwidth:"+node.clientWidth);
       //lists: go through listelements li.
       //if listelements are longer than 30% of page its horizontal
       var lis = node.getElementsByTagName("li");
-      console.log("blocks: lis.length" +lis.length);
-      console.log("blocks: lis[0].length"+lis[0].innerHTML.length);
+      //console.log("blocks: lis.length" +lis.length);
+      //console.log("blocks: lis[0].length"+lis[0].innerHTML.length);
       var maxw = 0;
       for(var l=0;l<lis.length;l++){
         if(lis[l].innerHTML.length>maxw)maxw = lis[l].innerHTML.length;
       }
       var screenw = document.getElementsByTagName("body")[0].clientWidth; //pagenode.clientWidth;
       var maxChar = screenw / 8;
-      console.log("maxChar dynamic:"+maxChar);
+      //console.log("maxChar dynamic:"+maxChar);
       maxChar = maxChar * 0.5; //50%
       if(maxChar===0)maxChar = 20;
-      console.log("maxChar:"+maxChar);
+      //console.log("maxChar:"+maxChar);
       if(maxw<maxChar && lis.length>3){
         node.classList.add("vertical");
       }else{
@@ -373,7 +373,7 @@ newtheme.addBlockClassesToElements = function(gridcontainer){
       node.areaname="footer";
       node.style.gridArea = "footer";
     }
-    console.log("blocks:classlist of element"+node.classList);
+    //console.log("blocks:classlist of element"+node.classList);
     if(node.classList.contains("listblock")){
       //tread listblocks always as vertical? just try it:
       node.classList.add("vertical");